필요 없는 파일 제거 + 이동
