--- GloomyDiary/Base/BaseView.swift
@@ -1,28 +1,0 @@
-//
-//  BaseView.swift
-//  GloomyDiary
-//
-//  Created by 디해 on 8/5/24.
-//
-
-import UIKit
-
-class BaseView: UIView {
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        
-        setup()
-        addSubviews()
-        setupConstraints()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    func setup() { }
-    
-    func addSubviews() { }
-    
-    func setupConstraints() { }
-}
