+++ GloomyDiary/Base/BaseViewController.swift
@@ -8,23 +8,19 @@
 import UIKit
 import RxSwift
 
-class BaseViewController<T: BaseView>: UIViewController {
+class BaseViewController<T: UIView>: UIViewController {
     var contentView: T
     
-    init(_ contentView: T = T(), logID: String) {
+    init(_ contentView: T = T()) {
         self.contentView = contentView
         super.init(nibName: nil, bundle: nil)
-    }
-    
-    override func loadView() {
-        self.view = contentView
     }
     
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
+    override func loadView() {
+        self.view = contentView
     }
 }
