--- GloomyDiary/Base/Protocol/FromTabSwitchable.swift
@@ -1,12 +1,0 @@
-//
-//  FromTabSwitchable.swift
-//  GloomyDiary
-//
-//  Created by 디해 on 10/26/24.
-//
-
-import UIKit
-
-protocol FromTabSwitchable: UIViewController {
-    func playTabDisappearingAnimation() async
-}
