--- GloomyDiary/Base/Protocol/ToTabSwitchable.swift
@@ -1,12 +1,0 @@
-//
-//  ToTabSwitchable.swift
-//  GloomyDiary
-//
-//  Created by 디해 on 10/26/24.
-//
-
-import UIKit
-
-protocol ToTabSwitchable: UIViewController {
-    func playTabAppearingAnimation() async
-}
